+++ android/src/main/java/com/bitmovin/player/reactnative/PlayerModule.kt
@@ -6,7 +6,9 @@ import com.facebook.react.bridge.*
 import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.uimanager.UIManagerModule
 
-@ReactModule(name = PlayerModule.name)
+private const val MODULE_NAME = "PlayerModule"
+
+@ReactModule(name = MODULE_NAME)
 class PlayerModule(private val context: ReactApplicationContext) : ReactContextBaseJavaModule(context) {
     /**
      * In-memory mapping from `nativeId`s to `Player` instances.
@@ -16,10 +18,7 @@ class PlayerModule(private val context: ReactApplicationContext) : ReactContextB
     /**
      * JS exported module name.
      */
-    companion object {
-        const val name = "PlayerModule"
-    }
-    override fun getName() = PlayerModule.name
+    override fun getName() = MODULE_NAME
 
     /**
      * Fetches the `Player` instance associated with `nativeId` from the internal players.
@@ -277,6 +276,43 @@ class PlayerModule(private val context: ReactApplicationContext) : ReactContextB
                 }
             }
             promise.resolve(subtitleTracks)
+        }
+    }
+
+    /**
+     * Schedules an `AdItem` in the `nativeId`'s associated player.
+     * @param nativeId Target player id.
+     * @param adItemJson Json representation of the `AdItem` to be scheduled.
+     */
+    @ReactMethod
+    fun scheduleAd(nativeId: NativeId, adItemJson: ReadableMap?) {
+        JsonConverter.toAdItem(adItemJson)?.let { adItem ->
+            uiManager()?.addUIBlock {
+                players[nativeId]?.scheduleAd(adItem)
+            }
+        }
+    }
+
+    /**
+     * Skips the current ad in `nativeId`'s associated player.
+     * Has no effect if the current ad is not skippable or if no ad is being played back.
+     * @param nativeId Target player id.
+     */
+    @ReactMethod
+    fun skipAd(nativeId: NativeId) {
+        uiManager()?.addUIBlock {
+            players[nativeId]?.skipAd()
+        }
+    }
+
+    /**
+     * Returns `true` while an ad is being played back or when main content playback has been paused for ad playback.
+     * @param nativeId Target player id.
+     */
+    @ReactMethod
+    fun isAd(nativeId: NativeId, promise: Promise) {
+        uiManager()?.addUIBlock {
+            promise.resolve(players[nativeId]?.isAd)
         }
     }