+++ src/components/PlayerView/events.ts
@@ -1,21 +1,39 @@
 import { NativeSyntheticEvent } from 'react-native';
 import {
+  AdBreakFinishedEvent,
+  AdBreakStartedEvent,
+  AdClickedEvent,
+  AdErrorEvent,
+  AdFinishedEvent,
+  AdManifestLoadedEvent,
+  AdManifestLoadEvent,
+  AdQuartileEvent,
+  AdScheduledEvent,
+  AdSkippedEvent,
+  AdStartedEvent,
   DestroyEvent,
   Event,
   MutedEvent,
   PausedEvent,
-  PlayEvent,
+  PictureInPictureAvailabilityChangedEvent,
+  PictureInPictureEnterEvent,
+  PictureInPictureEnteredEvent,
+  PictureInPictureExitEvent,
+  PictureInPictureExitedEvent,
   PlaybackFinishedEvent,
   PlayerActiveEvent,
   PlayerErrorEvent,
   PlayerWarningEvent,
+  PlayEvent,
   PlayingEvent,
   ReadyEvent,
-  SeekEvent,
   SeekedEvent,
+  SeekEvent,
+  StallStartedEvent,
+  StallEndedEvent,
   SourceErrorEvent,
-  SourceLoadEvent,
   SourceLoadedEvent,
+  SourceLoadEvent,
   SourceUnloadedEvent,
   SourceWarningEvent,
   SubtitleAddedEvent,
@@ -30,10 +48,26 @@ import {
  * Used to generate the specific events interface for each component.
  */
 interface EventProps {
+  onAdBreakFinished: AdBreakFinishedEvent;
+  onAdBreakStarted: AdBreakStartedEvent;
+  onAdClicked: AdClickedEvent;
+  onAdError: AdErrorEvent;
+  onAdFinished: AdFinishedEvent;
+  onAdManifestLoad: AdManifestLoadEvent;
+  onAdManifestLoaded: AdManifestLoadedEvent;
+  onAdQuartile: AdQuartileEvent;
+  onAdScheduled: AdScheduledEvent;
+  onAdSkipped: AdSkippedEvent;
+  onAdStarted: AdStartedEvent;
   onDestroy: DestroyEvent;
   onEvent: Event;
   onMuted: MutedEvent;
   onPaused: PausedEvent;
+  onPictureInPictureAvailabilityChanged: PictureInPictureAvailabilityChangedEvent;
+  onPictureInPictureEnter: PictureInPictureEnterEvent;
+  onPictureInPictureEntered: PictureInPictureEnteredEvent;
+  onPictureInPictureExit: PictureInPictureExitEvent;
+  onPictureInPictureExited: PictureInPictureExitedEvent;
   onPlay: PlayEvent;
   onPlaybackFinished: PlaybackFinishedEvent;
   onPlayerActive: PlayerActiveEvent;
@@ -43,6 +77,8 @@ interface EventProps {
   onReady: ReadyEvent;
   onSeek: SeekEvent;
   onSeeked: SeekedEvent;
+  onStallStarted: StallStartedEvent;
+  onStallEnded: StallEndedEvent;
   onSourceError: SourceErrorEvent;
   onSourceLoad: SourceLoadEvent;
   onSourceLoaded: SourceLoadedEvent;