+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.4.0] (2022-11-24)
+
+This version introduces the advertising API, full Picture in Picture support and exposes Stall events.
+
+### Added
+
+- Advertising API support.
+- Complete Picture in Picture support.
+- `onStallStarted`/`onStallEnded` events support. (Thanks to @joornby-angel)
+
 ## [0.3.1] (2022-10-26)
 
 Adds tweaks configuration support.
@@ -67,6 +77,7 @@ No support for custom UI yet.
 - Support for listening most of `Player` and `Source` events via `PlayerView`'s component props.
 - Simple React Native app to exemplify and test library features in development.
 
+[0.4.0]: https://github.com/bitmovin/bitmovin-player-react-native/releases/tag/v0.4.0
 [0.3.1]: https://github.com/bitmovin/bitmovin-player-react-native/releases/tag/v0.3.1
 [0.3.0]: https://github.com/bitmovin/bitmovin-player-react-native/releases/tag/v0.3.0
 [0.2.1]: https://github.com/bitmovin/bitmovin-player-react-native/releases/tag/v0.2.1